requirements.txt
attrs==23.1.0
Brotli==1.0.9
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==2.0.4
cryptography==41.0.3
exceptiongroup==1.1.3
h11==0.14.0
idna==3.4
lxml==4.9.3
numpy==1.26.0
outcome==1.3.0.post0
Pillow==10.1.0
pip==23.3
pycparser==2.21
pyOpenSSL==23.2.0
PySocks==1.7.1
python-docx==1.1.0
requests==2.31.0
selenium==4.15.2
setuptools==68.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.23.1
trio-websocket==0.11.1
typing_extensions==4.8.0
urllib3==1.26.18
wheel==0.41.2
wsproto==1.2.0
